backend/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
SQLAlchemy==2.0.21
python-dotenv==1.0.0
tensorflow==2.13.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
shap==0.42.1
lime>=0.2.0.1
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
kaleido>=0.2.1
Pillow>=10.0.0
base64
io
Werkzeug==2.3.7
pytest==7.4.2
pytest-flask==1.2.0
gunicorn==21.2.0
joblib==1.3.2
